refactor planning exports helpers
- reduce code duplication
- use event dates in planning advisories

SDBELGA-1067
